+++ lib/src/argument_parser.dart
@@ -1,8 +1,10 @@
 import 'package:args/args.dart';
+import 'package:faster_driver/src/shards.dart';
 import 'package:faster_driver/src/utils/colorize/colorizing.dart';
 
 class ArgumentParser {
   static const separator = ',,,';
+
   Args parse(List<String> args) {
     final separated = args
         .map((a) => a.replaceAll(separator, ' ')) //
@@ -25,6 +27,10 @@ class ArgumentParser {
       testArguments: params.wasParsed(_testArgumentsArg)
           ? _splitTestArguments(params[_testArgumentsArg])
           : [],
+      shard: Shard.fromString(
+        totalShards: params[_totalShardsArg],
+        shardIndex: params[_shardIndexArg],
+      ),
     );
   }
 
@@ -66,6 +72,7 @@ class ArgumentParser {
 
 class Split {
   const Split(this.value, {required this.contained});
+
   final String value;
   final bool contained;
 }
@@ -77,12 +84,14 @@ class MainArgs implements Args {
     required this.directory,
     required this.file,
     required this.testArguments,
+    required this.shard,
     required this.template,
   });
 
   final String directory;
   final String file;
   final List<String> testArguments;
+  final Shard? shard;
   final String? template;
 }
 
@@ -105,11 +114,25 @@ class ArgumentException implements Exception {
   ArgumentException.unknownArguments(List<String> args, FormatException e)
       : message = 'Unknown arguments in $args\n$e';
 
+  ArgumentException.shardMissingShardArgument()
+      : message = 'You have only passed one shard argument, you need to pass '
+            'both --$_totalShardsArg and --$_shardIndexArg';
+
+  ArgumentException.shardInvalidTotal(int shards)
+      : message = 'Total shards argument has to be 1 or greater, '
+            'the current is $shards';
+
+  ArgumentException.shardOutOfBounds({required int shards, required int index})
+      : message = 'The index of a shard ($index) it out of bounds. '
+            'Should be between 0 and ${shards - 1}';
+
   final String message;
 }
 
 const _fileArg = 'file';
 const _templateArg = 'template';
+const _totalShardsArg = 'total-shards';
+const _shardIndexArg = 'shard-index';
 const _helpArg = 'help';
 const _testArgumentsArg = 'test-args';
 ArgParser _scriptParameters = ArgParser()
@@ -126,6 +149,18 @@ ArgParser _scriptParameters = ArgParser()
   ..addOption(
     _templateArg,
     help: 'Template for aggregated test file',
+  )
+  ..addOption(
+    _totalShardsArg,
+    help: 'Tests can be sharded with the "--total-shards" and "--shard-index" '
+        'arguments, allowing you to split up your test suites '
+        'and run them separately.',
+  )
+  ..addOption(
+    _shardIndexArg,
+    help: 'Tests can be sharded with the "--total-shards" and "--shard-index" '
+        'arguments, allowing you to split up your test suites '
+        'and run them separately.',
   )
   ..addFlag(
     _helpArg,